+++ src/pretalx/orga/forms/event.py
@@ -6,11 +6,13 @@ from django.core.exceptions import ValidationError
 from django.utils.timezone import now
 from django.utils.translation import ugettext_lazy as _
 from hierarkey.forms import HierarkeyForm
+from i18nfield.fields import I18nFormField, I18nTextarea
 from i18nfield.forms import I18nFormMixin, I18nModelForm
 
 from pretalx.common.css import validate_css
 from pretalx.common.forms.fields import IMAGE_EXTENSIONS, ExtensionFileField
 from pretalx.common.mixins.forms import ReadOnlyFlag
+from pretalx.common.phrases import phrases
 from pretalx.event.models.event import Event, Event_SettingsStore
 from pretalx.orga.forms.widgets import HeaderSelect, MultipleLanguagesWidget
 
@@ -281,3 +283,68 @@ class MailSettingsForm(ReadOnlyFlag, I18nFormMixin, HierarkeyForm):
                     'Your administrator can add an instance-wide bypass. If you use this bypass, please also adjust your Privacy Policy.'
                 )
             )
+
+
+class ReviewSettingsForm(ReadOnlyFlag, I18nFormMixin, HierarkeyForm):
+    review_deadline = forms.DateTimeField(label=_('Review deadline'), required=False)
+    allow_override_votes = forms.BooleanField(
+        label=_('Allow override votes'),
+        help_text=_(
+            'Review teams can be assigned a fixed amount of "override votes": Positive or negative vetos each reviewer can assign.'
+        ),
+        required=False,
+    )
+    review_score_mandatory = forms.BooleanField(
+        label=_('Require a review score'), required=False
+    )
+    review_text_mandatory = forms.BooleanField(
+        label=_('Require a review text'), required=False
+    )
+    review_help_text = I18nFormField(
+        label=_('Help text for reviewers'),
+        help_text=_(
+            'This text will be shown at the top of every review, as long as reviews can be created or edited.'
+        )
+        + ' '
+        + phrases.base.use_markdown,
+        widget=I18nTextarea,
+        required=False,
+    )
+    review_min_score = forms.IntegerField(
+        label=_('Minimum score'), help_text=_('The minimum score reviewers can assign')
+    )
+    review_max_score = forms.IntegerField(
+        label=_('Maximum score'), help_text=_('The maximum score reviewers can assign')
+    )
+
+    def __init__(self, obj, *args, **kwargs):
+        super().__init__(*args, obj=obj, **kwargs)
+        if getattr(obj, 'slug'):
+            additional = _(
+                'You can configure override votes <a href="{link}">in the team settings</a>.'
+            ).format(link=obj.orga_urls.team_settings)
+            self.fields['allow_override_votes'].help_text += f' {additional}'
+        minimum = int(obj.settings.review_min_score)
+        maximum = int(obj.settings.review_max_score)
+        self.fields['review_deadline'].widget = forms.DateTimeInput(
+            attrs={'class': 'datetimepickerfield'}
+        )
+        for number in range(abs(maximum - minimum + 1)):
+            index = minimum + number
+            self.fields[f'review_score_name_{index}'] = forms.CharField(
+                label=_('Score label ({})').format(index),
+                help_text=_(
+                    'Human readable explanation of what a score of "{}" actually means, e.g. "great!".'
+                ).format(index),
+                required=False,
+            )
+
+    def clean(self):
+        data = self.cleaned_data
+        minimum = int(data.get('review_min_score'))
+        maximum = int(data.get('review_max_score'))
+        if minimum >= maximum:
+            raise forms.ValidationError(
+                _('Please assign a minimum score smaller than the maximum score!')
+            )
+        return data